requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.1
django-rest-passwordreset==1.4.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
install==1.3.5
mysqlclient==2.2.4
numpy==1.24.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
pytesseract==0.3.10
pytz @ file:///C:/b/abs_6ap4tsz1ox/croot/pytz_1713974360290/work
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
